web pages

+++ src/main/java/picstorage/rest/PicturesController.java
@@ -2,16 +2,22 @@ package picstorage.rest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.util.FileCopyUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import picstorage.domain.BytePicture;
+import picstorage.domain.Comment;
 import picstorage.domain.Picture;
 import picstorage.services.PictureService;
 import picstorage.services.UserInfoService;
 
+import javax.imageio.ImageIO;
 import javax.servlet.http.HttpServletResponse;
+import java.awt.*;
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.Date;
 import java.util.List;
@@ -57,8 +63,12 @@ public class PicturesController {
     @RequestMapping(value="/{id}", method= RequestMethod.DELETE)
     @ResponseBody
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void updatePictureInfo(@PathVariable long id) {
-        pictureService.deletePicture(id);
+    public void deletePicture(@PathVariable long id) {
+        Picture pic = pictureService.getPicture(id);
+        String userName = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (pic.getCreator().getLogin().equals(userName)) {
+            pictureService.deletePicture(id);
+        }
     }
 
     @RequestMapping(value="/{id}/pic", method= RequestMethod.GET)
@@ -72,22 +82,51 @@ public class PicturesController {
         }
     }
 
+    @RequestMapping(value="/{id}/thumb", method= RequestMethod.GET)
+    public void handleThumbDownload(@PathVariable long id, HttpServletResponse response) {
+        BytePicture bytePicture = pictureService.getBytePicture(id);
+        try {
+            FileCopyUtils.copy(bytePicture.getThumbnail(), response.getOutputStream());
+            response.flushBuffer();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
     @RequestMapping(method=RequestMethod.POST)
     @ResponseBody
     @ResponseStatus(HttpStatus.CREATED)
-    public Picture handleFileUpload(@RequestBody Picture picture,
-                                                 @RequestParam("file") MultipartFile file){
+    public Picture handleFileUpload(@RequestParam("title") String title,
+                                    @RequestParam("description") String description,
+                                    @RequestParam("file") MultipartFile file){
+        String userName = SecurityContextHolder.getContext().getAuthentication().getName();
+        Picture picture = new Picture(userInfoService.getUserInfo(userName), new Date());
+        picture.setTitle(title);
+        picture.setDescription(description);
         if (!file.isEmpty()) {
             try {
-
-                return pictureService.addPicture(
-                        new Picture(userInfoService.getUserInfo("test"), new Date()),
-                        new BytePicture(file.getOriginalFilename(), file.getBytes()));
+                ByteArrayOutputStream os = new ByteArrayOutputStream();
+                BufferedImage img = new BufferedImage(75, 75, BufferedImage.TYPE_INT_RGB);
+                img.createGraphics().drawImage(ImageIO.read(file.getInputStream())
+                        .getScaledInstance(75, 75, Image.SCALE_SMOOTH),0,0,null);
+                ImageIO.write(img, "jpg", os);
+                return pictureService.addPicture(picture,
+                        new BytePicture(file.getOriginalFilename(), file.getBytes(), os.toByteArray()));
             } catch (Exception e) {
                 return picture;
             }
         } else {
             return picture;
         }
+    }
+
+    @ResponseBody
+    @RequestMapping(value="/{id}/addcomment", method= RequestMethod.POST)
+    public Picture addComment(@PathVariable long id, @RequestParam("text") String text) {
+        String userName = SecurityContextHolder.getContext().getAuthentication().getName();
+        return pictureService.addComment(id, new Comment(
+                userInfoService.getUserInfo(userName),
+                new Date(),
+                text));
     }
 }
